Fix errors in sdp_utils_tests.cpp

+++ libs/Microsoft.MixedReality.WebRTC.Native/test/sdp_utils_tests.cpp
@@ -156,8 +156,8 @@ struct RaiiBuffer {
 // Check mrsSdpForceCodecs() forces the audio codec, without adding an
 // unsupported video codec.
 TEST(SdpUtils, ForceCodecs) {
-  size_t len = sizeof(kSdpFullString) * 2;
-  RaiiBuffer buffer(len);
+  uint64_t len = sizeof(kSdpFullString) * 2;
+  RaiiBuffer buffer((size_t)len);
   ASSERT_NE(nullptr, buffer._data);
   // Force audio to "opus" only. Don't change video because "h264" is not
   // advertized as supported in the input message
@@ -166,25 +166,25 @@ TEST(SdpUtils, ForceCodecs) {
   mrsSdpForceCodecs(kSdpFullString, audio_filter, video_filter, buffer._data,
                     &len);
   ASSERT_EQ(sizeof(kSdpForcedAudioOpus), len);
-  ASSERT_EQ(0, memcmp(kSdpForcedAudioOpus, buffer._data, len));
+  ASSERT_EQ(0, memcmp(kSdpForcedAudioOpus, buffer._data, (size_t)len));
 }
 
 // No-op if codecs are not supported
 TEST(SdpUtils, ForceCodecsNotSupported) {
-  size_t len = sizeof(kSdpFullString) * 2;
-  RaiiBuffer buffer(len);
+  uint64_t len = sizeof(kSdpFullString) * 2;
+  RaiiBuffer buffer((size_t)len);
   ASSERT_NE(nullptr, buffer._data);
   SdpFilter audio_filter{"random_non_existing_audio_codec", ""};
   SdpFilter video_filter{"random_non_existing_video_codec", ""};
   mrsSdpForceCodecs(kSdpFullString, audio_filter, video_filter, buffer._data,
                     &len);
   ASSERT_EQ(sizeof(kSdpFullString), len);
-  ASSERT_EQ(0, memcmp(kSdpFullString, buffer._data, len));
+  ASSERT_EQ(0, memcmp(kSdpFullString, buffer._data, (size_t)len));
 }
 
 // Buffer too small
 TEST(SdpUtils, ForceCodecsShortBuffer) {
-  size_t len = 32;  // too short on purpose
+  uint64_t len = 32;  // too short on purpose
   char buffer[32];
   SdpFilter audio_filter{"opus", ""};
   SdpFilter video_filter{"h264", ""};
